moving modules to modules/ folder
